% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{adwin_func}
\alias{adwin_func}
\title{ADWIN algorithm}
\usage{
adwin_func(
  x,
  parameters = list(delta = 0.002, clock = 32, max_buckets = 5, min_window_length = 5,
    grace_period = 10)
)
}
\arguments{
\item{x}{univariate series}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
ADWIN algorithm
}
