% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithm_parameters.R
\name{algorithm_name_to_function}
\alias{algorithm_name_to_function}
\title{Internal function for mapping a change point algorithm to its related function}
\usage{
algorithm_name_to_function(algorithm_name)
}
\arguments{
\item{algorithm_name}{string within `all_sota_algorithms()`}
}
\value{
the related function to apply the get the binarized change point results
}
\description{
Internal function for mapping a change point algorithm to its related function
}
