% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithm_parameters.R
\name{algorithm_name_to_parameters}
\alias{algorithm_name_to_parameters}
\title{Internal function for mapping a change point algorithm to its grid of parameters}
\usage{
algorithm_name_to_parameters(algorithm_name)
}
\arguments{
\item{algorithm_name}{string within `all_sota_algorithms()`}
}
\value{
the selected grid for this algorithm
}
\description{
Internal function for mapping a change point algorithm to its grid of parameters
}
