% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{all_cost_functions}
\alias{all_cost_functions}
\title{All cost functions used in the pipeline}
\usage{
all_cost_functions()
}
\value{
Vector of all exported functions
}
\description{
All cost functions used in the pipeline
}
