% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithm_parameters.R
\name{all_sota_algorithms}
\alias{all_sota_algorithms}
\title{List of selected state-of-the-art change point algorithms}
\usage{
all_sota_algorithms()
}
\value{
a string vector listing the different selected sota algorithms
}
\description{
List of selected state-of-the-art change point algorithms
}
