% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_output_plots.R
\name{average_rank_best_single}
\alias{average_rank_best_single}
\title{Prepared CD data frame ready for plotting, for the single-best setting}
\usage{
average_rank_best_single(OUTPUT_DATA_FOLDER)
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
data frame ready for plotting for the single-best setting
}
\description{
Prepared CD data frame ready for plotting, for the single-best setting
}
