% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{breakfast_func}
\alias{breakfast_func}
\title{breakfast algorithms}
\usage{
breakfast_func(
  x,
  parameters = list(scoring = "wbs", M = 1000, selection = "sdll")
)
}
\arguments{
\item{x}{univariate series}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
breakfast algorithms
}
