% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_scoring_algorithm.R
\name{build_df_df2}
\alias{build_df_df2}
\title{Compute the cost of each interval (df) and of consecutive pairs of interval (df2)}
\usage{
build_df_df2(x, t, cost_func)
}
\arguments{
\item{x}{univariate or multivariate vector series, on which change points need to be found}

\item{t}{a subset of 1:(length(x)+1) in the increasing order, typically
an interval s:e or 1:(length(x)+1), but should be numeric}

\item{cost_func}{Selected cost function}
}
\value{
a list of two elements, each being a data frame.
The first two columns of each data frame is an interval [s, e) (s included and e excluded)
The third column is the cost of this interval
}
\description{
Compute the cost of each interval (df) and of consecutive pairs of interval (df2)
}
