% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{changeforest_func}
\alias{changeforest_func}
\title{changeforest algorithm}
\usage{
changeforest_func(
  df,
  parameters = list(method = "random_forest", segmentation_type = "bs")
)
}
\arguments{
\item{df}{data frame with one or multiple columns}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
changeforest algorithm
}
