% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_scoring_algorithm.R
\name{compute_diff_costs}
\alias{compute_diff_costs}
\title{Compute the difference of cost after removing one of the remaining change points,
for each possible change point that can be removed. All the costs are global
to the whole interval of interest c(l$df$s[1], l$df$e[nrow(l$df)])}
\usage{
compute_diff_costs(l)
}
\arguments{
\item{l}{list of (df,df2) data frame}
}
\value{
difference of cost after removing each change point. This is the
difference of global costs.

For a specific index, if the difference is large and positive,
the cost is increasing by removing this specific change point,
so this change point is important and should not be removed.

For a specific index, if the difference is close to zero or negative
the difference of cost is small by removing this specific change point,
so this change point is not important and can be removed.
}
\description{
Compute the difference of cost after removing one of the remaining change points,
for each possible change point that can be removed. All the costs are global
to the whole interval of interest c(l$df$s[1], l$df$e[nrow(l$df)])
}
