% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{compute_median_user_groundtruth}
\alias{compute_median_user_groundtruth}
\title{Compute the median ground truth user given the labels given by multiple users}
\usage{
compute_median_user_groundtruth(df)
}
\arguments{
\item{df}{containing the ground truth of each user (columns) for each data
set (rows)}
}
\value{
a list (one element per data set) of the median ground truth user
}
\description{
Compute the median ground truth user given the labels given by multiple users
}
