% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_selection.R
\name{convert_partitions_to_data_frame}
\alias{convert_partitions_to_data_frame}
\title{Convert the output obtained from scoth_recursive_partitions to a data.frame with
the selection/binarization at each level for the method}
\usage{
convert_partitions_to_data_frame(threshold, raw_output, scoring)
}
\arguments{
\item{threshold}{numeric selected threshold to be indicated in `selection_threshold_included` column}

\item{raw_output}{list with partitions and changepoint_gains at each level obtained
from the `scoth_recursive_partitions` function}

\item{scoring}{initial scoring related to the changepoint_gains, only used for checking}
}
\value{
data.frame with columns `selection_threshold_included`, `selection_level`,
`selection`, `selection_additional`, for a single threshold and at all levels, giving the
binarization for this threshold
}
\description{
Convert the output obtained from scoth_recursive_partitions to a data.frame with
the selection/binarization at each level for the method
}
