% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{cpnp_func}
\alias{cpnp_func}
\title{Non parametric change point algorithm}
\usage{
cpnp_func(x, parameters = list(pen = "SIC", Q = 10))
}
\arguments{
\item{x}{univariate series}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
Non parametric change point algorithm
}
