% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_selection.R
\name{default_single_row_output}
\alias{default_single_row_output}
\title{Helper for extreme thresholds functions (`multithreshold_default_inf` and `multithreshold_default_min`)}
\usage{
default_single_row_output(
  single_threshold,
  single_prediction,
  index,
  levels = NA
)
}
\arguments{
\item{single_threshold}{threshold either default_score or +Inf, corresponding to an extreme threshold}

\item{single_prediction}{prediction that all the index elements should have (0 for +Inf, or 1 for default_score threshold)}

\item{index}{index of the scoring of the form 2:n with n the number of row of the related scoring}

\item{levels}{vector of levels of the form 1:nb_levels}
}
\value{
a pre-filled template of the selection output row for this threshold, with the selection
filled and the selection_additional template
}
\description{
Helper for extreme thresholds functions (`multithreshold_default_inf` and `multithreshold_default_min`)
}
