% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_results.R
\name{expand_grid_df}
\alias{expand_grid_df}
\title{Expand grid helper for `build_grid`}
\usage{
expand_grid_df(...)
}
\arguments{
\item{...}{Arguments, each of them being a numeric or string vector (no list),
e.g. `...` is `p=c(1,2),method=c("a","b")`}
}
\value{
Data frame output of the grid of possible combinations of the output,
names with the name of the original arguments, and with the increment begining
by the last columns. Please use `build_grid` for change point detection, this
is only an internal helper
}
\description{
Expand grid helper for `build_grid`
}
