% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_F1.R
\name{f1_func}
\alias{f1_func}
\title{Compute the F1-score given precision and recall, handling all the corner cases}
\usage{
f1_func(precision, recall)
}
\arguments{
\item{precision}{numeric value}

\item{recall}{numeric value}

\item{W}{window range for considering the detected change point as valid, w.r.t. the ground truth indexes}
}
\value{
the F1-score
}
\description{
Compute the F1-score given precision and recall, handling all the corner cases
}
