% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_selection_algorithm.R
\name{find_indexes}
\alias{find_indexes}
\title{Select change points}
\usage{
find_indexes(changepoint_gain, s = 1, threshold = 0.03)
}
\arguments{
\item{changepoint_gain}{vector of length the number of possible change points, with the gain
w.r.t to the global cost of [s, e). The larger gain the more probable it is a change}

\item{s}{start of the interval of interest [s, e)}

\item{threshold}{within [0,1]}
}
\value{
vector of indexes (w.r.t. absolute index 1, not to s) considered as change points
}
\description{
Select change points
}
