% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{get_oracle}
\alias{get_oracle}
\title{Function for reproducing the results using our methodology in the oracle setting for a single metric}
\usage{
get_oracle(metric = "F1", OUTPUT_DATA_FOLDER, full_results = FALSE)
}
\arguments{
\item{metric}{string representing the metrics of interest, within: "F1", "F1_biased" or "cover"}

\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
the average scores and deviations for a single metric
}
\description{
Function for reproducing the results using our methodology in the oracle setting for a single metric
}
