% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_output_tables.R
\name{get_oracle_sota}
\alias{get_oracle_sota}
\title{Function for reproducing the results using for state-of-the-art algorithms
in the oracle setting for a single evaluation metric}
\usage{
get_oracle_sota(
  metric = "F1",
  OUTPUT_DATA_FOLDER,
  full_results = FALSE,
  mapping_sota = mapping_family
)
}
\arguments{
\item{metric}{string representing the metrics of interest, within: "F1", "F1_biased" or "cover"}

\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}

\item{full_results}{whether to output the full results or the summarized table.
With FALSE: outputs the summarized table that averages the
F1 score over the datasets. With TRUE: output the data frame with the F1 score
for each dataset}

\item{mapping_sota}{function mapping the algorithm_fullname to its family.
With `NULL`: keep the original algorithm_name (corresponding to the package from
which the method has been computed). With `mapping_detailed`: separate the
different search methods, resulting in 15+ rows; With `mapping_family`: group
the algorithm based on their family (selection path, penalty, misc)}
}
\value{
a data frame, with the average scores and deviations for each method,
using the oracle setting (algorithm parameters reaching the best performance
for each dataset), and for a single evaluation metric
}
\description{
Function for reproducing the results using for state-of-the-art algorithms
in the oracle setting for a single evaluation metric
}
