% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{get_oracle_table}
\alias{get_oracle_table}
\title{Function for reproducing the results using our methodology in the oracle setting}
\usage{
get_oracle_table(OUTPUT_DATA_FOLDER)
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
the average scores and deviations for the three metrics
}
\description{
Function for reproducing the results using our methodology in the oracle setting
}
