% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_output_tables.R
\name{get_oracle_table_sota}
\alias{get_oracle_table_sota}
\title{Function for reproducing the results using for state-of-the-art algorithms
in the oracle setting}
\usage{
get_oracle_table_sota(OUTPUT_DATA_FOLDER, mapping_sota = mapping_family)
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}

\item{mapping_sota}{function mapping the algorithm_fullname to its family.
With `NULL`: keep the original algorithm_name (corresponding to the package from
which the method has been computed). With `mapping_detailed`: separate the
different search methods, resulting in 15+ rows; With `mapping_family`: group
the algorithm based on their family (chain of subsets, selection path, penalty,
or misc)}
}
\value{
a data frame, with the average scores and deviations for each method,
using the oracle setting (algorithm parameters reaching the best performance
for each dataset), and for the three evaluation metrics
}
\description{
Function for reproducing the results using for state-of-the-art algorithms
in the oracle setting
}
