% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_selection_algorithm.R
\name{get_partition_from_changepoint_gain}
\alias{get_partition_from_changepoint_gain}
\title{Build partition from selected change points based on a criteria}
\usage{
get_partition_from_changepoint_gain(
  changepoint_gain,
  x,
  cost_func,
  s,
  e,
  threshold
)
}
\arguments{
\item{changepoint_gain}{vector of length the number of possible change points, with the gain
w.r.t to the global cost of [s, e). The larger gain the more probable it is a change}

\item{x}{univariate (vector) or multivariate (data.frame) series, on which change points need to be found}

\item{cost_func}{a cost function}

\item{s}{start of the interval of interest [s, e)}

\item{e}{end of the interval of interest [s, e) (w.r.t. absolute index 1)}

\item{threshold}{within [0,1]}
}
\value{
data frame, each row being an interval [s, e) between two change points (or the border)
with the associated cost
}
\description{
Build partition from selected change points based on a criteria
}
