% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_selection_algorithm.R
\name{get_partition_from_idx}
\alias{get_partition_from_idx}
\title{Convert the selected index change points to a series of ranges indicating each interval}
\usage{
get_partition_from_idx(idx = c(), x, cost_func, s = 1, e = length(x) + 1)
}
\arguments{
\item{idx}{vector of indexes considered as change points (w.r.t. absolute index 1)}

\item{x}{univariate (vector) or multivariate (data.frame) series, on which change points need to be found}

\item{cost_func}{a cost function}

\item{s}{start of the interval of interest [s, e) (w.r.t. absolute index 1)}

\item{e}{end of the interval of interest [s, e) (w.r.t. absolute index 1)}
}
\value{
data frame, each row being an interval [s, e) between two change points (or the border)
}
\description{
Convert the selected index change points to a series of ranges indicating each interval
}
