% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{get_pen.value}
\alias{get_pen.value}
\title{Helper function for `killick_func` and `cpnp_func` functions
Retrieve the penalty value given the input}
\usage{
get_pen.value(pen)
}
\arguments{
\item{pen}{string with the form of a penalty value (e.g., "pen0.1"), or of
a penalty method (e.g., "BIC")}
}
\value{
extract the penalty scheme (e.g., "0.1*log(n)") if the input is a
penalty value, or put 0 if the input is a penalty method
}
\description{
Helper function for `killick_func` and `cpnp_func` functions
Retrieve the penalty value given the input
}
