% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{get_single_best_table}
\alias{get_single_best_table}
\title{Function for reproducing the results using our methodology in the single-best setting}
\usage{
get_single_best_table(OUTPUT_DATA_FOLDER, full_results = FALSE)
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
the average score and deviation, along with the selected best setting
}
\description{
Function for reproducing the results using our methodology in the single-best setting
}
