% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_output_tables.R
\name{get_single_best_table_sota}
\alias{get_single_best_table_sota}
\title{Function for reproducing the results using the for state-of-the-art algorithms
in the single-best setting}
\usage{
get_single_best_table_sota(
  OUTPUT_DATA_FOLDER,
  full_results = FALSE,
  mapping_sota = mapping_family
)
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}

\item{full_results}{whether to output the full results or the summarized table.
With FALSE: outputs the summarized table that averages the
F1 score over the datasets. With TRUE: output the data frame with the F1 score
for each dataset}

\item{mapping_sota}{function mapping the algorithm_fullname to its family.
With `NULL`: keep the original algorithm_name (corresponding to the package from
which the method has been computed). With `mapping_detailed`: separate the
different search methods, resulting in 15+ rows; With `mapping_family`: group
the algorithm based on their family (selection path, penalty, misc)}
}
\value{
a data frame, with the average scores and deviations for each method,
using the single best setting for each algorithm
}
\description{
Function for reproducing the results using the for state-of-the-art algorithms
in the single-best setting
}
