% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_selection_algorithm.R
\name{iterate_changepoint_gain_from_partition}
\alias{iterate_changepoint_gain_from_partition}
\title{Update to the lower granularity changepoint_gain without the need to recompute all the elements}
\usage{
iterate_changepoint_gain_from_partition(
  partition,
  changepoint_gain,
  cost_initial
)
}
\arguments{
\item{partition}{data frame, each row being an interval [s, e) between two change points and the
related cost. Those are the iterated ones}

\item{changepoint_gain}{vector of length the number of possible change points, with the gain
w.r.t to the *global* cost of the whole series. The larger gain the more probable it is a change
Here it is always `changepoint_gain` of the *global* cost, not the next ones}

\item{cost_initial}{cost of the whole series, used for normalizing the gain to
a new partition}
}
\value{
vector of length the number of possible change points, with the gain
w.r.t to the local cost of each partition [s, e), and where the remaining elements
corresponding to previous change points assigned at larger granularity are set to NA
}
\description{
Update to the lower granularity changepoint_gain without the need to recompute all the elements
}
