% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{killick_func}
\alias{killick_func}
\title{PELT and related change point algorithms}
\usage{
killick_func(
  x,
  parameters = list(cost = "meanvar", search = "AMOC", pen = "SIC")
)
}
\arguments{
\item{x}{univariate series}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
PELT and related change point algorithms
}
