% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{lasso_func}
\alias{lasso_func}
\title{Lasso-based change point algorithms}
\usage{
lasso_func(
  x,
  parameters = list(lambda = 1, gamma = 0, eps = 0.001, trend = TRUE)
)
}
\arguments{
\item{x}{univariate series}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
Lasso-based change point algorithms
}
