% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_groundtruths}
\alias{load_groundtruths}
\title{Main function for loading change point ground truths
Load the ground truth with one row per series}
\usage{
load_groundtruths(PUBLIC_DATA_FOLDER = "~/anomaly-detection/scoth/public-data")
}
\arguments{
\item{PUBLIC_DATA_FOLDER}{filepath to the public-data folder}
}
\value{
Data frame with each row being a node (shanghai_license, well_log...) and
each column being a different user labeling each series. The users are user_X with X
a certain number. The first column `groundtruth` is, for each series, one of the user
that has the best Jaccard with the other users (~ the "median" user of each series).
Each cell is the index of change point selected by this user on this series
- If the cell is NULL, the user did not label this series,
- If the cell is integer(0), the user decided that no change point is present
- If the cell is a vector of index, those are the corresponding change point indexes
as hinted by the user.
The smallest possible index is 2 (meaning that a change point has happened at 2)
}
\description{
Main function for loading change point ground truths
Load the ground truth with one row per series
}
