% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_kpis}
\alias{load_kpis}
\title{Main function for loading data set KPIs. Load as a data frame with one row per series}
\usage{
load_kpis(PUBLIC_DATA_FOLDER = "~/anomaly-detection/scoth/public-data")
}
\arguments{
\item{PUBLIC_DATA_FOLDER}{filepath to the public-data folder}
}
\value{
a data frame with two columns:
- the `dataset_name`,
- a `kpis` data frame with as number of columns the number of features
}
\description{
Main function for loading data set KPIs. Load as a data frame with one row per series
}
