% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{load_summary_func_initial}
\alias{load_summary_func_initial}
\title{Helper for loading the summarized results (obtained from the `pipeline_summarization_func` function)}
\usage{
load_summary_func_initial(OUTPUT_DATA_FOLDER, metric = "F1")
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}

\item{metric}{string representing the metrics of interest, either "F1" or "TCPD" (for F1-biased and cover)}
}
\value{
the data frame giving the metrics (F1 scores and additional metrics) for
each dataset, cost method, threshold, level, and ground truth user
}
\description{
Helper for loading the summarized results (obtained from the `pipeline_summarization_func` function)
}
