% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_F1.R
\name{metrics_F1_func}
\alias{metrics_F1_func}
\title{Compute the F1 score metric}
\usage{
metrics_F1_func(
  binarization = data.frame(index = 2:101, predictions = rep(0, 100), groundtruth =
    rep(0, 100))
)
}
\arguments{
\item{binarization}{data frame with three columns:
- index: the index of the possible change points from 2 to T
- predictions: binary vector, with 1 when the index is detected as a change point
- groundtruth: binary vector, with 1 when the index is set by the ground truth as a change point}
}
\value{
a list of the F1 score, precision, recall, TP (true positives), along with
the number of ground truth change points and of predicted change points
}
\description{
Compute the F1 score metric
}
