% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_F1.R
\name{metrics_F1_TP_func}
\alias{metrics_F1_TP_func}
\title{Extract the number of true positives (TP) given the predicted and ground truth
indexes, for a certain window. We follow here the recommendations from the
following paper `https://ceur-ws.org/Vol-1226/paper31.pdf`}
\usage{
metrics_F1_TP_func(
  pred_idx = c(1000, 1001, 1015),
  gt_idx = c(1002, 1018, 1030),
  W = 5
)
}
\arguments{
\item{pred_idx}{vector of predicted change point indexes}

\item{gt_idx}{vector of ground truth change point indexes}

\item{W}{window range for considering the detected change point as valid, w.r.t. the ground truth indexes}
}
\value{
the number of true positives
}
\description{
Extract the number of true positives (TP) given the predicted and ground truth
indexes, for a certain window. We follow here the recommendations from the
following paper `https://ceur-ws.org/Vol-1226/paper31.pdf`
}
