% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_TCPD.R
\name{metrics_TCPD_func}
\alias{metrics_TCPD_func}
\title{Compute the F1-biased metric and the cover metric following the TCPD benchmark,
reproducing the code available at the following address:
`https://github.com/alan-turing-institute/TCPDBench/blob/master/analysis/scripts/metrics.py`
The corresponding tests are available in the test folder.}
\usage{
metrics_TCPD_func(binarization)
}
\arguments{
\item{binarization}{data frame with three columns:
- index: the index of the possible change points from 2 to T
- predictions: binary vector, with 1 when the index is detected as a change point
- groundtruth: binary vector, with 1 when the index is set by the ground truth as a change point}
}
\value{
a list of the F1 biased score, precision, recall, covering score,
additional recall parts (for each user), covering parts (for each user), along with
the number of ground truth change points and of predicted change points
}
\description{
Compute the F1-biased metric and the cover metric following the TCPD benchmark,
reproducing the code available at the following address:
`https://github.com/alan-turing-institute/TCPDBench/blob/master/analysis/scripts/metrics.py`
The corresponding tests are available in the test folder.
}
