% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_selection.R
\name{multithreshold_default_inf}
\alias{multithreshold_default_inf}
\title{Selection output data.frame row for the +Inf threshold in the case
of multithreshold selection method}
\usage{
multithreshold_default_inf(scoring, nb_levels, cost_initial)
}
\arguments{
\item{scoring}{data frame with columns index and score}

\item{nb_levels}{number of recursive levels}

\item{cost_initial}{initial cost of the whole data set series}
}
\value{
selection output rows for the +Inf threshold as a data.frame
}
\description{
Selection output data.frame row for the +Inf threshold in the case
of multithreshold selection method
}
