% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_selection.R
\name{multithreshold_default_min}
\alias{multithreshold_default_min}
\title{Selection output data.frame row for the default_score threshold in the case
of multithreshold selection method}
\usage{
multithreshold_default_min(scoring, nb_levels, default_score)
}
\arguments{
\item{scoring}{data frame with columns index and score}

\item{nb_levels}{number of recursive levels}

\item{default_score}{minimum possible scoring for the scoring score}
}
\value{
selection output rows for the default_score threshold as a data.frame
}
\description{
Selection output data.frame row for the default_score threshold in the case
of multithreshold selection method
}
