% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_results.R
\name{multivariate_func}
\alias{multivariate_func}
\title{Convert a univariate algorithm to the multivariate setting}
\usage{
multivariate_func(f)
}
\arguments{
\item{f}{change point detection algorithm giving as input a univariate series
and a list of parameters, and outputting a data.frame with index and predictions}
}
\value{
function that returns the aggregated data.frame of the predictions
obtained by applying f on each univariate series
}
\description{
This function is used for the algorithms that can originally only run on
univariate series. The final predictions are obtained by taking the union of
all the predictions made (as done by PELT in the changepoint package)
}
