% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{normalize_multivariate_series}
\alias{normalize_multivariate_series}
\title{Normalization of the multivariate series}
\usage{
normalize_multivariate_series(x_cur)
}
\arguments{
\item{x_cur}{Multivariate data frame series, that needs to be normalized among features}
}
\value{
the normalized data frame
}
\description{
Normalization of the multivariate series
}
