% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{ocp_func}
\alias{ocp_func}
\title{OCP change point algorithm}
\usage{
ocp_func(df, parameters = list(l = 10, a = 0.01, b = 0.01, k = 0.01))
}
\arguments{
\item{df}{data frame with one or multiple columns}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
OCP change point algorithm
}
