% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_evaluation.R
\name{pipeline_evaluation_single_selection}
\alias{pipeline_evaluation_single_selection}
\title{Compute the evaluation of a single data set}
\usage{
pipeline_evaluation_single_selection(df, groundtruth_node, metric = "F1")
}
\arguments{
\item{df}{data frame with the *selection* obtained for each dataset, cost, threshold and level}

\item{groundtruth_node}{ground truth row giving the selected ground truth
indexes for each user for this data set}

\item{metric}{string representing the metrics of interest, either "F1" or "TCPD" (for F1-biased and cover)}
}
\value{
a data frame giving, for each selection and each ground truth user,
the *evaluation* (list of metrics for each ground truth user), along with
additional information
}
\description{
Compute the evaluation of a single data set
}
