% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_scoring.R
\name{pipeline_scoring_func}
\alias{pipeline_scoring_func}
\title{Compute scorings of all the data sets and save them on disk}
\usage{
pipeline_scoring_func(loaded_kpis, cost_methods, OUTPUT_DATA_FOLDER)
}
\arguments{
\item{loaded_kpis}{data frame containing the `kpis` series (each univariate or multivariate)
and the corresponding `dataset_name`}

\item{cost_methods}{vector of cost functions names among the available cost functions}

\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
save a RDS file for the scoring results of each data set, each file
being a data frame with `length(cost_methods)` rows, each row containing the
*scoring* (data frame with the score obtained for each possible change point index),
along additional information
}
\description{
Compute scorings of all the data sets and save them on disk
}
