% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_scoring.R
\name{pipeline_scoring_series}
\alias{pipeline_scoring_series}
\title{Compute scoring of a single data set (univariate or multivariate)}
\usage{
pipeline_scoring_series(x, cost_methods = c("reg_cusum_L2", "cusum_L2"))
}
\arguments{
\item{x}{univariate (vector) or multivariate (data.frame) series}

\item{cost_methods}{vector of cost functions names among the available cost functions}
}
\value{
a data frame with `length(cost_methods)` rows, each row containing the
*scoring* (data frame with the score obtained for each possible change point index),
along with the `cost_method`, the `cost_initial` (initial cost of the whole series),
the `scoring_time_taken` (in seconds) and whether the `scoring_has_error`
}
\description{
Compute scoring of a single data set (univariate or multivariate)
}
