% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_selection.R
\name{pipeline_selection_func}
\alias{pipeline_selection_func}
\title{Compute selections of each scoring and save them on disk}
\usage{
pipeline_selection_func(
  thresholds = c(seq(from = 0.03, to = 0.2, by = 0.01), 0.3, 0.4, 0.5, 0.6, 0.7, 1),
  nb_levels = 10,
  OUTPUT_DATA_FOLDER
)
}
\arguments{
\item{thresholds}{vector of selected thresholds}

\item{nb_levels}{number of recursive levels}

\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
save a RDS file for the selection results of each data set. Each file
is a data frame. For each cost_method, threshold, and level, get the
corresponding *selection* (data frame with the 0/1 result obtained for each
possible change point index), along with additional information
}
\description{
After loading the scorings of each dataset, apply the selection thresholds and
iterate over the levels, before finally saving each RDS (one for each dataset)
inside the OUTPUT_DATA_FOLDER
}
