% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_results.R
\name{pipeline_selection_func_sota}
\alias{pipeline_selection_func_sota}
\title{Compute binarized selections for a given algorithm for state-of-the-art algorithms
and save them on disk}
\usage{
pipeline_selection_func_sota(algorithm_name, loaded_kpis, OUTPUT_DATA_FOLDER)
}
\arguments{
\item{algorithm_name}{change point algorithm name}

\item{loaded_kpis}{data frame containing the `kpis` series (each univariate or multivariate)
and the corresponding `dataset_name`}

\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
save a RDS file for the selection results of each data set. For each
parameter of the algorithm, get the corresponding *selection* (data frame with
the 0/1 result obtained for each possible change point index)
}
\description{
Compute binarized selections for a given algorithm for state-of-the-art algorithms
and save them on disk
}
