% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_selection.R
\name{pipeline_selection_single_post_process}
\alias{pipeline_selection_single_post_process}
\title{Helper for post-processing the `output_list`}
\usage{
pipeline_selection_single_post_process(
  output_list,
  scoring,
  nb_levels,
  cost_initial,
  default_score,
  x
)
}
\arguments{
\item{output_list}{list with the possible thresholds at each level up to a certain level}

\item{scoring}{data frame with columns index and score}

\item{nb_levels}{number of levels considered}

\item{cost_initial}{initial cost of the whole series}

\item{default_score}{minimum possible scoring for the scoring score}

\item{x}{the current univariate or multivariate data set series}
}
\value{
post-processed `output_list` with the added thresholds `0` (default scores) and `+Inf`,
added to `output_list`
}
\description{
Helper for post-processing the `output_list`
}
