% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_selection.R
\name{pipeline_selection_single_scoring}
\alias{pipeline_selection_single_scoring}
\title{Compute the selection of a single data set and scoring over all thresholds and levels}
\usage{
pipeline_selection_single_scoring(
  scoring,
  thresholds,
  nb_levels,
  cost_initial,
  cost_func,
  x
)
}
\arguments{
\item{scoring}{data frame with columns index and score}

\item{thresholds}{vector of selected thresholds}

\item{nb_levels}{number of recursive levels}

\item{cost_initial}{initial cost of the whole data set series}

\item{cost_func}{current cost function}

\item{x}{the current univariate or multivariate data set series}
}
\value{
a data frame giving, for each threshold and level, the corresponding
*selection* (data frame with the 0/1 result obtained for each possible change
point index), along with additional information
}
\description{
Compute the selection of a single data set and scoring over all thresholds and levels
}
