% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_summarization.R
\name{pipeline_summarization_func}
\alias{pipeline_summarization_func}
\title{Summarize all the evaluation results into a single file, for each metric}
\usage{
pipeline_summarization_func(metric, OUTPUT_DATA_FOLDER)
}
\arguments{
\item{metric}{string representing the metrics of interest, either "F1" or "TCPD" (for F1-biased and cover)}

\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
save a RDS file for the summarized results. Each file
is a data frame. For each dataset_name, cost_method, threshold, level, and
ground truth user, report the scores (F1 score, F1-biased score or cover)
}
\description{
Summarize all the evaluation results into a single file, for each metric
}
