% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_output_plots.R
\name{plot_cd_figures}
\alias{plot_cd_figures}
\title{Function for reproducing the CD plot figures of the PAKDD'25 paper}
\usage{
plot_cd_figures(OUTPUT_DATA_FOLDER)
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}
}
\value{
nothing, only output the two CD plots as pdfs
}
\description{
Function for reproducing the CD plot figures of the PAKDD'25 paper
}
