% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plots.R
\name{plot_figure_qualitative_results}
\alias{plot_figure_qualitative_results}
\title{Function for reproducing the figure of the PAKDD'25 paper regarding the qualitative results}
\usage{
plot_figure_qualitative_results(
  OUTPUT_DATA_FOLDER,
  groundtruth,
  outfile = file.path(OUTPUT_DATA_FOLDER, "figure_qualitative_results.pdf"),
  width = 9,
  height = 6
)
}
\arguments{
\item{OUTPUT_DATA_FOLDER}{filepath to the output data folder}

\item{groundtruth}{data frame with ground truth indexes obtained with `load_groundtruths`}

\item{outfile}{filepath to the output plot}

\item{width}{plot size (width)}

\item{height}{plot size (height)}
}
\value{
nothing, only output the figure to the `outfile` file path, as a pdf
}
\description{
Function for reproducing the figure of the PAKDD'25 paper regarding the qualitative results
}
