% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{rfpop_func}
\alias{rfpop_func}
\title{RFPOP change point algorithm}
\usage{
rfpop_func(x, parameters = list(cost = "Outlier", pen = "pen0.3"))
}
\arguments{
\item{x}{univariate series}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
RFPOP change point algorithm
}
