% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_scoring_algorithm.R
\name{scoth_scoring_func}
\alias{scoth_scoring_func}
\title{Main scoring function: cost for each possible change point along with the initial cost of the series}
\usage{
scoth_scoring_func(x, cost_func, verbose = TRUE)
}
\arguments{
\item{x}{univariate (vector) or multivariate (data.frame) series, on which change points need to be found}

\item{cost_func}{a cost function name}

\item{verbose}{whether to print advance of the algorithm}
}
\value{
a list with the initial cost (a numeric) and the "score" for each possible
change point (first element = first possible change point i.e. {1} alone and {2...} after)
}
\description{
Main scoring function: cost for each possible change point along with the initial cost of the series
}
