% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_summarization.R
\name{summarize_current_experiment_evaluation}
\alias{summarize_current_experiment_evaluation}
\title{Summarize a single dataset evaluation result}
\usage{
summarize_current_experiment_evaluation(df)
}
\arguments{
\item{df}{data frame with the *evaluation* obtained for each dataset, cost, threshold, level}
}
\value{
the same evaluation with the ground truth user added as an additional column
}
\description{
Summarize a single dataset evaluation result
}
