% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoth_scoring_algorithm.R
\name{update_costs}
\alias{update_costs}
\title{Remove the `i`th change point and update the costs}
\usage{
update_costs(l, i, x, cost_func)
}
\arguments{
\item{l}{list of (df,df2) data frame}

\item{i}{change point `i`th the be removed, i.e. the end time of the row
`i` of df, which is also the start time of the row `i+1` of df. `i` should
be in 1:k with k=nrow(l$df)-1 the number of change points in `l`}

\item{x}{univariate or multivariate vector series, on which change points need to be found}

\item{cost_func}{a cost function}
}
\value{
updated `l` list of (df,df2) data frame
}
\description{
Remove the `i`th change point and update the costs
}
