% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota_algorithms.R
\name{zero_func}
\alias{zero_func}
\title{Zero change point algorithm (predict 0 change points)}
\usage{
zero_func(x, parameters = list())
}
\arguments{
\item{x}{univariate series}

\item{parameters}{list of parameters for this algorithm}
}
\value{
data frame with the binarized prediction for each possible change point index
}
\description{
Zero change point algorithm (predict 0 change points)
}
